requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
djangorestframework==3.15.2
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-decouple==3.8
requests==2.32.3
rsa==4.9
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
